+++ app/app.go
@@ -2,5 +2,5 @@ package app
 
 const (
 	Name    = "circumflex"
-	Version = "2.2-dev"
+	Version = "2.2"
 )